+++ brent-measures_uk.csv
@@ -1,3 +1,1 @@
 #time,date,susceptible,exposed,infectious,recovered,dead,immune,num infections today,num hospitalisations today,hospital bed occupancy,num hospitalisations today (data)
-1,2/3/2020,323978,3749,2303,933,0,0,0,0,20,0.0
-2,3/3/2020,322491,4448,2791,1233,0,0,0,0,25,0.0
