+++ facs/base/utils.py
@@ -17,7 +17,7 @@ def probability(prob):
     return np.random.random() < prob
 
 
-def get_random_int(high):
+def get_random_int(high) -> int:
     """Return a random integer between 0 and high."""
     return np.random.randint(0, high)
 
@@ -75,3 +75,11 @@ def log_recovery(t, x, y, age, rank):
 
     print("{},{},{},{}".format(t, x, y, age), file=out_inf, flush=True)
     num_recoveries_today += 1
+
+
+def calc_dist(x1, y1, x2, y2):
+    return (np.abs(x1 - x2) ** 2 + np.abs(y1 - y2) ** 2) ** 0.5
+
+
+def calc_dist_cheap(x1, y1, x2, y2):
+    return np.abs(x1 - x2) + np.abs(y1 - y2)
